DI Framework integration / Tests implemented/ Test scenario 7Docummentation Folder indexed
Insertion of Zenject plugin into project.
Scene test with an  audio manager and isolation of dependencies of Mono framework into DI to make movements.
Tests with NUnit inserted on the project to verify the inputs and serve as example to the next code with will arrive.
Docs folder to LDD / MDD / GDD / And any kind of code implementation